Take the gossip sender from the envelope, not the connection
Gossip relays a payload through a self-stream, whose loopback connection
reports the local peer at both ends. Both auth decisions read that peer,
so a relayed payload was verified against this node's own key and handed
the owner's private-route access - and nothing signs the destination, so
the publisher chose the handler every subscriber ran. PRIVATE_POST_BLOCK
and PRIVATE_POST_GATEWAY_SETTINGS were reachable on every node following
the topic.

PublishUpdateToFollowers already stamps the author's node id and Publish
signs the envelope with that node's key, so the proof of authorship is
there: read the sender from the envelope on a self-stream and verify
against it. Naming another node gains nothing, since the signature is
then checked against that node's key. The stream passed on to the handler
reports the sender that was verified.

Delivery gets its own route and handler beside the timeline it writes to,
instead of reusing the owner's compose route, and takes a tweet only from
its author's own node.

TestGossip_SelfPublishResignsAndForwards asserted a re-sign that 64c51b80
removed and has been failing since; it now pins that the sender's own
signature is forwarded untouched.
