+++ core/handler/tweet.go
@@ -29,7 +29,6 @@ package handler
 
 import (
 	"context"
-	"crypto/ed25519"
 	"errors"
 	"fmt"
 	"strings"
@@ -103,77 +102,9 @@ type TweetNotifier interface {
 	Add(not domain.Notification) error
 }
 
-func StreamTimelineTweetHandler(
-	authRepo OwnerTweetStorer,
-	tweetRepo TweetsStorer,
-	timelineRepo TimelineUpdater,
-	followRepo TweetFollowChecker,
-	userRepo TweetUserFetcher,
-) warpnet.WarpHandlerFunc {
-	return func(buf []byte, _ warpnet.WarpStream) (any, error) {
-		var ev event.NewTweetEvent
-		if err := json.Unmarshal(buf, &ev); err != nil {
-			return nil, err
-		}
-		if ev.Moderation != nil && !ev.Moderation.IsOk {
-			return nil, tweetRepo.Blocklist(ev.Id)
-		}
-		if err := validateTweetEvent(ev); err != nil {
-			return nil, err
-		}
-
-		owner := authRepo.GetOwner()
-		if owner.UserId == ev.UserId {
-			return event.Accepted, nil
-		}
-		if followRepo == nil || !followRepo.IsFollowing(owner.UserId, ev.UserId) {
-			return event.Accepted, nil
-		}
-		pubKey, err := authorPubKey(ev.UserId, userRepo)
-		if err != nil {
-			log.Warnf("timeline: dropping tweet from %s: %v", ev.UserId, err)
-			return nil, err
-		}
-		if err := ev.Verify(pubKey); err != nil {
-			log.Warnf("timeline: dropping tweet from %s: signature invalid: %v", ev.UserId, err)
-			return nil, ErrForeignTweetAuthor
-		}
-
-		tweet, err := tweetRepo.Create(ev.UserId, ev)
-		if err != nil {
-			return nil, err
-		}
-		if tweet.Id == "" {
-			return tweet, warpnet.WarpError("timeline handler: empty tweet id")
-		}
-		if err := timelineRepo.AddTweetToTimeline(owner.UserId, tweet); err != nil {
-			log.Infof("fail adding tweet to timeline: %v", err)
-		}
-		return tweet, nil
-	}
-}
-
-func authorPubKey(authorId string, userRepo TweetUserFetcher) (ed25519.PublicKey, error) {
-	if userRepo == nil {
-		return nil, ErrForeignTweetAuthor
-	}
-	author, err := userRepo.Get(authorId)
-	if err != nil || author.NodeId == "" {
-		return nil, ErrForeignTweetAuthor
-	}
-	pubKey := warpnet.FromIDToPubKey(warpnet.FromStringToPeerID(author.NodeId))
-	if len(pubKey) != ed25519.PublicKeySize {
-		return nil, ErrForeignTweetAuthor
-	}
-	return pubKey, nil
-}
-
-const ErrForeignTweetAuthor = warpnet.WarpError("tweet: author does not live on the sending node")
-
 func StreamNewTweetHandler(
 	broadcaster TweetBroadcaster,
 	authRepo OwnerTweetStorer,
-	privKey ed25519.PrivateKey,
 	tweetRepo TweetsStorer,
 	timelineRepo TimelineUpdater,
 	followRepo TweetFollowChecker,
@@ -240,7 +171,7 @@ func StreamNewTweetHandler(
 				ImageKeys: tweet.ImageKeys,
 				VideoKey:  tweet.VideoKey,
 				Poll:      tweet.Poll,
-			}.Signed(privKey)
+			}
 			bt, _ := json.Marshal(respTweetEvent)
 			if err := broadcaster.PublishUpdateToFollowers(owner.UserId, event.PUBLIC_POST_TIMELINE, bt); err != nil {
 				log.Errorf("broadcaster publish owner tweet update: %v", err)
